proj_spam_detector/requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==3.0.4
cloudpickle==1.6.0
decorator==4.4.2
Django==3.1.4
djangorestframework==3.12.2
gunicorn==19.6.0
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.0.0
jupyter-client==6.1.7
jupyter-core==4.7.0
numpy==1.19.2
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
python-dateutil==2.8.1
pytz==2020.5
pyzmq==20.0.0
requests==2.25.1
scikit-learn==0.23.2
scipy==1.6.0
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
nltk==3.5
